create and get

+++ src/main/java/co/com/cesde/tienda/modelo/ProductoModelo.java
@@ -36,7 +36,7 @@ public class ProductoModelo {
         return cantidad;
     }
 
-    public void setCantidad(float cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
 
@@ -44,7 +44,7 @@ public class ProductoModelo {
         return precio;
     }
 
-    public void setPrecio(float precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
 }
